pyproject.toml
[project]
name = "questvoyager"
version = "0.1.0"
description = ""
authors = [
    {name = "Jeongho-Shin",email = "purplecollar@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "pandas (>=2.2.3,<3.0.0)",
    "streamlit (>=1.44.0,<2.0.0)",
    "streamlit-option-menu (>=0.4.0,<0.5.0)",
    "python-docx (>=1.1.2,<2.0.0)",
    "pymupdf (>=1.25.5,<2.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "openai (>=1.70.0,<2.0.0)",
    "langchain (>=0.3.22,<0.4.0)",
    "langchain-openai (>=0.3.11,<0.4.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "litellm (>=1.65.1,<2.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
